#30daysToCode #day11 Refactor actions, actionTypes, reducers to index
