+++ src/actions/index.js
@@ -1,0 +1,3 @@
+import PlayerActionCreators from './player';
+
+export { PlayerActionCreators };
