+++ src/actiontypes/index.js
@@ -1,0 +1,2 @@
+import * as PlayerActionTypes from './player';
+export { PlayerActionTypes };
