+++ src/actiontypes/player.js
