+++ src/index.js
@@ -2,14 +2,14 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { createStore } from 'redux';
-import PlayerReducer from 'reducers/player';
+import reducers from 'reducers';
 import 'css/bootstrap.min.css';
 import 'css/index.css';
 import App from 'containers/App/App';
 import registerServiceWorker from './registerServiceWorker';
 
 const store = createStore(
-    PlayerReducer,
+    reducers,
     window.devToolsExtension && window.devToolsExtension(),
 );
