+++ src/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import PlayerReducers from './player';
+
+const rootReducer = combineReducers({
+    player: PlayerReducers,
+});
+
+export default rootReducer;
